Bindings
* Simplify name from CnbBindings (a convention that makes my eyes bleed) so
that you aren't stuttering.  I'm still a bit undecided on the naming of the
modules themselves as well.

* Documentation, documentation, documentation!

* Lots of the final keyword.  Prevents unanticipated modification via extension.

* Constructor/Method overloading allows you to isolate the behaviors that are
really hard to test.  Note that I've isolated everything that needs to read
from the environment and feel confident that the methods that include it are
so trivial they're not worth testing.  Some overloads are non-public to enable
testing but not later usage.

* Since this is a public API, I'm being really verbose with the nullability
annotations.  Most of the time you wouldn't use them, by they do make IDE
users' lives better.

Signed-off-by: Ben Hale <bhale@vmware.com>