BindingsEnvironmentPostProcessor
The previous implementation was *so* complex!  I realize that you grabbed it
from the previous implementation, but you ran into one of our biggest
failings, too much complexity, too early.  Among other things the looping
priority was inverted (looping on the wrong thing).  This simplifies the whole
into a handful of lines, and once again notice the judicious use of
constructors to make it testable.

Signed-off-by: Ben Hale <bhale@vmware.com>