Isolate Boot-specific Processors
Previously, the Boot-specific processors were in the same package as the core
types.  This minimzed the visibility of some constructors used for testing,
but as the list of processors has grown (and will continue to grow), the core
types were being overwhelmed.

This change moves the Boot-specific processors out to an additional package
where they can be easily identified.

Signed-off-by: Ben Hale <bhale@vmware.com>