Deferred Logging
Previously the logging put out by the library happened synchronously with the
code executing.  Because this library typically runs so early in the Spring
Boot lifecycle that logging isn't yet configured, it needs to defer that
logging until later.

This change converts the library to use Boot's DeferredLogger and listen for
an ApplicationPrepared event before logging out its messages.

Signed-off-by: Ben Hale <bhale@vmware.com>